requirements.txt
argparse
numpy
matplotlib
Bio 
pylab-sdk
pandas
scipy
